feat: #22 Review and creation of atomic components
