+++ src/components/common/footer.jsx
@@ -1,52 +1,12 @@
+import SocialNetworksComponent from "./footerComponents/socialNetworksComponent"
+import WebLinksComponent from "./footerComponents/webLinksComponent"
+
 function Footer() {
     return (
         <footer class="">
             <div class="d-flex flex-column flex-md-row align-content-md-center align-items-md-center justify-content-md-center pb-md-3 gap-md-5">
-                <ul class="order-1 order-md-3 d-flex flex-row justify-content-between px-3 pt-3">
-                    <li class="mx-md-1">
-                        <a href="" class="foot-opc fs-80">
-                            Nosotros
-                        </a>
-                    </li>
-                    <li class="mx-md-1">
-                        <a href="" class="foot-opc fs-80">
-                            FAQs
-                        </a>
-                    </li>
-                    <li class="mx-md-1">
-                        <a href="" class="foot-opc fs-80">
-                            Privacidad
-                        </a>
-                    </li>
-                    <li class="mx-md-1">
-                        <a href="" class="foot-opc fs-80">
-                            Términos
-                        </a>
-                    </li>
-                </ul>
-                <ul class="order-2 order-md-3 d-flex flex-row justify-content-center px-3 mt-3">
-                    <li class="mx-1">
-                        <img
-                            src="./imgs/rrss_fb.svg"
-                            alt=""
-                            class="img-foot img-fluid"
-                        />
-                    </li>
-                    <li class="mx-1">
-                        <img
-                            src="./imgs/rrss_ig.svg"
-                            alt=""
-                            class="img-foot img-fluid"
-                        />
-                    </li>
-                    <li class="mx-1">
-                        <img
-                            src="./imgs/rrss_mail.svg"
-                            alt=""
-                            class="img-foot img-fluid"
-                        />
-                    </li>
-                </ul>
+                <WebLinksComponent></WebLinksComponent>
+                <SocialNetworksComponent></SocialNetworksComponent>
                 <p class="order-3 order-md-1 mt-3 mb-0 pb-4 pb-md-0 text-center fs-80">
                     © 2022 LaRifamos
                 </p>
