+++ src/components/common/footerComponents/socialNetworksComponent.jsx
@@ -1,0 +1,30 @@
+function SocialNetworksComponent () {
+    return (
+        <ul class="order-2 order-md-3 d-flex flex-row justify-content-center px-3 mt-3">
+                    <li class="mx-1">
+                        <img
+                            src="./imgs/rrss_fb.svg"
+                            alt=""
+                            class="img-foot img-fluid"
+                        />
+                    </li>
+                    <li class="mx-1">
+                        <img
+                            src="./imgs/rrss_ig.svg"
+                            alt=""
+                            class="img-foot img-fluid"
+                        />
+                    </li>
+                    <li class="mx-1">
+                        <img
+                            src="./imgs/rrss_mail.svg"
+                            alt=""
+                            class="img-foot img-fluid"
+                        />
+                    </li>
+                </ul>
+    ); 
+
+}
+
+export default SocialNetworksComponent;
