+++ src/components/common/footerComponents/webLinksComponent.jsx
@@ -1,0 +1,28 @@
+function WebLinksComponent () {
+    return(
+        <ul class="order-1 order-md-3 d-flex flex-row justify-content-between px-3 pt-3">
+                    <li class="mx-md-1">
+                        <a href="" class="foot-opc fs-80">
+                            Nosotros
+                        </a>
+                    </li>
+                    <li class="mx-md-1">
+                        <a href="" class="foot-opc fs-80">
+                            FAQs
+                        </a>
+                    </li>
+                    <li class="mx-md-1">
+                        <a href="" class="foot-opc fs-80">
+                            Privacidad
+                        </a>
+                    </li>
+                    <li class="mx-md-1">
+                        <a href="" class="foot-opc fs-80">
+                            Términos
+                        </a>
+                    </li>
+                </ul>
+    ) 
+}
+
+export default WebLinksComponent;
