+++ src/components/common/navbarComponents/logoComponent.jsx
@@ -1,0 +1,26 @@
+function LogoComponent() {
+    return (
+        <div class="col-9 d-flex align-items-center mb-3">
+            <ul class="d-flex align-items-center">
+                <li class="short">
+                    <img
+                        class="logo"
+                        src="./imgs/Logo_LaRifamos.png"
+                        alt=""
+                    />
+                </li>
+                <li class="mx-2">
+                    <a
+                        href=""
+                        class="rifas-btn d-flex align-content-center justify-content-center"
+                    >
+                        <span>Rifas</span>
+                    </a>
+                </li>
+            </ul>
+        </div>
+
+    );
+}
+
+export default LogoComponent;
